+++ src/Symfony/Component/Messenger/Asynchronous/Routing/ContainerSenderLocator.php
@@ -1,0 +1,40 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Messenger\Asynchronous\Routing;
+
+use Psr\Container\ContainerInterface;
+use Symfony\Component\Messenger\Transport\SenderInterface;
+
+/**
+ * @author Samuel Roze <samuel.roze@gmail.com>
+ */
+class ContainerSenderLocator extends AbstractSenderLocator
+{
+    private $senderServiceLocator;
+    private $messageToSenderIdMapping;
+
+    public function __construct(ContainerInterface $senderServiceLocator, array $messageToSenderIdMapping)
+    {
+        $this->senderServiceLocator = $senderServiceLocator;
+        $this->messageToSenderIdMapping = $messageToSenderIdMapping;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getSenderForMessage($message): ?SenderInterface
+    {
+        $senderId = self::getValueFromMessageRouting($this->messageToSenderIdMapping, $message);
+
+        return $senderId ? $this->senderServiceLocator->get($senderId) : null;
+    }
+}