+++ src/Symfony/Component/Messenger/CHANGELOG.md
@@ -4,6 +4,9 @@ CHANGELOG
 4.2.0
 -----
 
+ * [BC BREAK] `SenderLocator` has been renamed to `ContainerSenderLocator`
+   Be careful as there is still a `SenderLocator` class, but it does not rely on a `ContainerInterface` to find senders.
+   Instead, it accepts the sender instance itself instead of its identifier in the container.
  * [BC BREAK] `MessageSubscriberInterface::getHandledMessages()` return value has changed. The value of an array item
    needs to be an associative array or the method name. 
  * `ValidationMiddleware::handle()` and `SendMessageMiddleware::handle()` now require an `Envelope` object